app/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
click==6.6
docopt==0.4.0
Flask==0.11.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.35
mandrill==1.0.57
MarkupSafe==0.23
pymongo==3.3.0
pytz==2016.6.1
PyYAML==3.12
redis==2.10.5
requests==2.11.1
Werkzeug==0.11.11
